+++ src/LightGraphs.jl
@@ -219,6 +219,7 @@ include("cycles/hawick-james.jl")
 include("cycles/karp.jl")
 include("cycles/basis.jl")
 include("cycles/limited_length.jl")
+include("traversals/traversals.jl")
 include("traversals/bfs.jl")
 include("traversals/bipartition.jl")
 include("traversals/greedy_color.jl")
@@ -229,6 +230,7 @@ include("traversals/diffusion.jl")
 include("connectivity.jl")
 include("distance.jl")
 include("edit_distance.jl")
+include("shortestpaths/shortestpaths.jl")
 include("shortestpaths/astar.jl")
 include("shortestpaths/bellman-ford.jl")
 include("shortestpaths/dijkstra.jl")
@@ -237,6 +239,13 @@ include("shortestpaths/desopo-pape.jl")
 include("shortestpaths/floyd-warshall.jl")
 include("shortestpaths/yen.jl")
 include("shortestpaths/spfa.jl")
+include("shortestpaths/deprecated/astar.jl")
+include("shortestpaths/deprecated/bellman-ford.jl")
+include("shortestpaths/deprecated/dijkstra.jl")
+include("shortestpaths/deprecated/johnson.jl")
+include("shortestpaths/deprecated/desopo-pape.jl")
+include("shortestpaths/deprecated/floyd-warshall.jl")
+include("shortestpaths/deprecated/spfa.jl")
 include("linalg/LinAlg.jl")
 include("operators.jl")
 include("persistence/common.jl")
