+++ src/deprecations.jl
@@ -6,3 +6,7 @@
 
 # Deprecated for more explicit function name. Can be removed for version 2.0.
 @deprecate saw self_avoiding_walk
+
+@deprecate a_star(g::AbstractGraph, s, t, distmx::AbstractMatrix, heuristic::Function) shortestpaths(g, s, t, distmx, AStar(heuristic=heuristic))
+@deprecate a_star(g::AbstractGraph, s, t, distmx::AbstractMatrix) shortestpaths(g, s, t, distmx, AStar())
+@deprecate a_star(g::AbstractGraph, s, t) shortestpaths(g, s, t, AStar())
