+++ src/shortestpaths/bellman-ford.jl
@@ -3,39 +3,33 @@
 
 # The Bellman Ford algorithm for single-source shortest path
 
-###################################################################
-#
-#   The type that capsulates the state of Bellman Ford algorithm
-#
-###################################################################
-using Base.Threads
-
 struct NegativeCycleError <: Exception end
 
-# AbstractPathState is defined in core
 """
-    BellmanFordState{T, U}
+    struct BellmanFord <: ShortestPathAlgorithm
 
-An `AbstractPathState` designed for Bellman-Ford shortest-paths calculations.
+The structure used to configure and specify that [`shortest_paths`](@ref)
+should use the [Bellman-Ford algorithm](http://en.wikipedia.org/wiki/Bellman–Ford_algorithm).
+No fields are specified or required.
+
+### Implementation Notes
+`BellmanFord` supports the following shortest-path functionality:
+- negative distance matrices / weights
+- (optional) multiple sources
+- all destinations
 """
-struct BellmanFordState{T<:Real, U<:Integer} <: AbstractPathState
+struct BellmanFord <: ShortestPathAlgorithm end
+struct BellmanFordResult{T, U<:Integer} <: ShortestPathResult
     parents::Vector{U}
     dists::Vector{T}
 end
 
-"""
-    bellman_ford_shortest_paths(g, s, distmx=weights(g))
-    bellman_ford_shortest_paths(g, ss, distmx=weights(g))
-
-Compute shortest paths between a source `s` (or list of sources `ss`) and all
-other nodes in graph `g` using the [Bellman-Ford algorithm](http://en.wikipedia.org/wiki/Bellman–Ford_algorithm).
-Return a [`LightGraphs.BellmanFordState`](@ref) with relevant traversal information.
-"""
-function bellman_ford_shortest_paths(
+function shortest_paths(
     graph::AbstractGraph{U},
     sources::AbstractVector{<:Integer},
-    distmx::AbstractMatrix{T}=weights(graph)
-    ) where T<:Real where U<:Integer
+    distmx::AbstractMatrix{T},
+    ::BellmanFord
+   ) where {T, U<:Integer}
 
     nvg = nv(graph)
     active = falses(nvg)
@@ -46,7 +40,7 @@ function bellman_ford_shortest_paths(
     no_changes = false
     new_active = falses(nvg)
 
-    for i in vertices(graph)
+    @inbounds for i in vertices(graph)
         no_changes = true
         new_active .= false
         for u in vertices(graph)[active]
@@ -60,77 +54,23 @@ function bellman_ford_shortest_paths(
                 end
             end
         end
-        if no_changes
-            break
-        end
+        no_changes && break
         active, new_active = new_active, active
     end
     no_changes || throw(NegativeCycleError())
-    return BellmanFordState(parents, dists)
+    return BellmanFordResult(parents, dists)
 end
 
-bellman_ford_shortest_paths(
-    graph::AbstractGraph{U},
-    v::Integer,
-    distmx::AbstractMatrix{T} = weights(graph);
-    ) where T<:Real where U<:Integer = bellman_ford_shortest_paths(graph, [v], distmx)
+shortest_paths(g::AbstractGraph, v::Integer, distmx::AbstractMatrix, alg::BellmanFord) =
+    shortest_paths(g, [v], distmx, alg)
 
-has_negative_edge_cycle(g::AbstractGraph) = false
+has_negative_weight_cycle(g::AbstractGraph, ::BellmanFord) = false
 
-function has_negative_edge_cycle(
-    g::AbstractGraph{U}, 
-    distmx::AbstractMatrix{T}
-    ) where T<:Real where U<:Integer
+function has_negative_weight_cycle(g::AbstractGraph, distmx::AbstractMatrix, alg::BellmanFord)
     try
-        bellman_ford_shortest_paths(g, vertices(g), distmx)
+        shortest_paths(g, vertices(g), distmx, alg)
     catch e
-        isa(e, NegativeCycleError) && return true
+        isa(e, ShortestPaths.NegativeCycleError) && return true
     end
     return false
 end
-
-"""
-    enumerate_paths(state[, vs])
-
-Given a path state `state` of type `AbstractPathState`, return a
-vector (indexed by vertex) of the paths between the source vertex used to
-compute the path state and a single destination vertex, a list of destination
-vertices, or the entire graph. For multiple destination vertices, each
-path is represented by a vector of vertices on the path between the source and
-the destination. Nonexistent paths will be indicated by an empty vector. For
-single destinations, the path is represented by a single vector of vertices,
-and will be length 0 if the path does not exist.
-
-### Implementation Notes
-For Floyd-Warshall path states, please note that the output is a bit different,
-since this algorithm calculates all shortest paths for all pairs of vertices:
-`enumerate_paths(state)` will return a vector (indexed by source vertex) of
-vectors (indexed by destination vertex) of paths. `enumerate_paths(state, v)`
-will return a vector (indexed by destination vertex) of paths from source `v`
-to all other vertices. In addition, `enumerate_paths(state, v, d)` will return
-a vector representing the path from vertex `v` to vertex `d`.
-"""
-function enumerate_paths(state::AbstractPathState, vs::Vector{<:Integer})
-    parents = state.parents
-    T = eltype(parents)
-
-    num_vs = length(vs)
-    all_paths = Vector{Vector{T}}(undef, num_vs)
-    for i = 1:num_vs
-        all_paths[i] = Vector{T}()
-        index = T(vs[i])
-        if parents[index] != 0 || parents[index] == index
-            while parents[index] != 0
-                push!(all_paths[i], index)
-                index = parents[index]
-            end
-            push!(all_paths[i], index)
-            reverse!(all_paths[i])
-        end
-    end
-    return all_paths
-end
-
-enumerate_paths(state::AbstractPathState, v) = enumerate_paths(state, [v])[1]
-enumerate_paths(state::AbstractPathState) = enumerate_paths(state, [1:length(state.parents);])
-
