+++ src/shortestpaths/desopo-pape.jl
@@ -1,48 +1,23 @@
 """
-    struct DEposoPapeState{T, U}
+    struct DEsopoPape <: ShortestPathAlgorithm
 
-An [`AbstractPathState`](@ref) designed for D`Esopo-Pape shortest-path calculations.
+The structure used to configure and specify that [`shortest_paths`](@ref)
+should use the [D'Esopo-Pape algorithm](http://web.mit.edu/dimitrib/www/SLF.pdf).
+No fields are specified or required.
+
+### Implementation Notes
+`DEsopoPape` supports the following shortest-path functionality:
+- non-negative distance matrices / weights
+- all destinations
 """
-struct DEsopoPapeState{T <:Real, U <: Integer} <: AbstractPathState
+struct DEsopoPape <: ShortestPathAlgorithm end
+struct DEsopoPapeResult{T, U<:Integer} <: ShortestPathResult
     parents::Vector{U}
     dists::Vector{T}
 end
 
-"""
-    desopo_pape_shortest_paths(g, src, distmx=weights(g))
-
-Compute shortest paths between a source `src` and all
-other nodes in graph `g` using the [D'Esopo-Pape algorithm](http://web.mit.edu/dimitrib/www/SLF.pdf).
-Return a [`LightGraphs.DEsopoPapeState`](@ref) with relevant traversal information.
-
-# Examples
-```jldoctest
-julia> using LightGraphs
-
-julia> ds = desopo_pape_shortest_paths(cycle_graph(5), 2);
-
-julia> ds.dists
-5-element Array{Int64,1}:
- 1
- 0
- 1
- 2
- 2
-
-julia> ds = desopo_pape_shortest_paths(path_graph(5), 2);
-
-julia> ds.dists
-5-element Array{Int64,1}:
- 1
- 0
- 1
- 2
- 3
-```
-"""
-function desopo_pape_shortest_paths(g::AbstractGraph, 
-    src::Integer,
-    distmx::AbstractMatrix{T} = weights(g)) where T <: Real
+function shortest_paths(g::AbstractGraph, src::Integer, distmx::AbstractMatrix, ::DEsopoPape)
+    T = eltype(distmx)
     U = eltype(g)
     nvg = nv(g)
     (src in 1:nvg) || throw(DomainError(src, "src should be in between 1 and $nvg"))
@@ -74,5 +49,5 @@ function desopo_pape_shortest_paths(g::AbstractGraph,
         end
     end
     
-    return DEsopoPapeState{T, U}(parents, dists)
+    return DEsopoPapeResult{T, U}(parents, dists)
 end
