+++ src/shortestpaths/johnson.jl
@@ -1,34 +1,30 @@
-
 """
-    struct JohnsonState{T, U}
+    struct Johnson <: ShortestPathAlgorithm
 
-An [`AbstractPathState`](@ref) designed for Johnson shortest-paths calculations.
-"""
-struct JohnsonState{T <: Real,U <: Integer} <: AbstractPathState
-    dists::Matrix{T}
-    parents::Matrix{U}
-end
+The structure used to configure and specify that [`shortest_paths`](@ref)
+should use the [Johnson algorithm](https://en.wikipedia.org/wiki/Johnson%27s_algorithm).
+No additional configuration parameters are specified or required.
 
-"""
-    johnson_shortest_paths(g, distmx=weights(g))
-
-Use the [Johnson algorithm](https://en.wikipedia.org/wiki/Johnson%27s_algorithm)
-to compute the shortest paths between all pairs of vertices in graph `g` using an
-optional distance matrix `distmx`.
-
-Return a [`LightGraphs.JohnsonState`](@ref) with relevant
-traversal information.
+### Implementation Notes
+`Johnson` supports the following shortest-path functionality:
+- non-negative distance matrices / weights
+- all-pairs shortest paths
 
 ### Performance
 Complexity: O(|V|*|E|)
 """
-function johnson_shortest_paths(g::AbstractGraph{U},
-    distmx::AbstractMatrix{T}=weights(g)) where T <: Real where U <: Integer
+struct Johnson <: ShortestPathAlgorithm end
 
+struct JohnsonResult{T, U<:Integer} <: ShortestPathResult
+    parents::Matrix{U}
+    dists::Matrix{T}
+end
+
+function shortest_paths(g::AbstractGraph{U}, distmx::AbstractMatrix{T}, ::Johnson) where {T, U<:Integer}
     nvg = nv(g)
     type_distmx = typeof(distmx)
     #Change when parallel implementation of Bellman Ford available
-    wt_transform = bellman_ford_shortest_paths(g, vertices(g), distmx).dists
+    wt_transform = LightGraphs.Experimental.ShortestPaths.dists(shortest_paths(g, vertices(g), distmx, BellmanFord()))
     
     if !type_distmx.mutable && type_distmx !=  LightGraphs.DefaultDistance
         distmx = sparse(distmx) #Change reference, not value
@@ -61,10 +57,12 @@ function johnson_shortest_paths(g::AbstractGraph{U},
         end
     end
 
-    return JohnsonState(dists, parents)
+    return JohnsonResult(parents, dists)
 end
 
-function enumerate_paths(s::JohnsonState{T,U}, v::Integer) where T <: Real where U <: Integer
+shortest_paths(g::AbstractGraph, alg::Johnson) = shortest_paths(g, weights(g), alg)
+
+function paths(s::JohnsonResult{T, U}, v::Integer) where {T, U <: Integer}
     pathinfo = s.parents[v, :]
     paths = Vector{Vector{U}}()
     for i in 1:length(pathinfo)
@@ -83,5 +81,5 @@ function enumerate_paths(s::JohnsonState{T,U}, v::Integer) where T <: Real where
     return paths
 end
 
-enumerate_paths(s::JohnsonState) = [enumerate_paths(s, v) for v in 1:size(s.parents, 1)]
-enumerate_paths(st::JohnsonState, s::Integer, d::Integer) = enumerate_paths(st, s)[d]
+paths(s::JohnsonResult) = [paths(s, v) for v in 1:size(s.parents, 1)]
+paths(st::JohnsonResult, s::Integer, d::Integer) = paths(st, s)[d]
