+++ src/traversals/bfs.jl
@@ -1,172 +1,94 @@
-### DEVELOPERS NOTE: BFS optimization experiments are typically
-### prototyped in gdistances!, since it's one of the simplest
-### BFS implementations.
+import Base.Sort, Base.Sort.Algorithm
+import Base:sort!
 
-"""
-    tree(parents)
+struct NOOPSortAlg <: Base.Sort.Algorithm end
+const NOOPSort = NOOPSortAlg()
 
-Convert a parents array into a directed graph.
-"""
-function tree(parents::AbstractVector{T}) where T <: Integer
-    n = T(length(parents))
-    t = DiGraph{T}(n)
-    for (v, u) in enumerate(parents)
-        if u > zero(T)  && u != v
-            add_edge!(t, u, v)
-        end
-    end
-    return t
+sort!(x, ::Integer, ::Integer, ::NOOPSortAlg, ::Base.Sort.Ordering) = x
+
+struct BFS{T<:Base.Sort.Algorithm} <: TraversalAlgorithm
+    sort_alg::T
 end
 
+BFS() = BFS(NOOPSort)
+
 """
-    bfs_parents(g, s[; dir=:out])
+    traverse_graph!(g, s, alg, state, neighborfn=outneighbors)
+    traverse_graph!(g, ss, alg, state, neighborfn=outneighbors)
 
-Perform a breadth-first search of graph `g` starting from vertex `s`.
-Return a vector of parent vertices indexed by vertex. If `dir` is specified,
-use the corresponding edge direction (`:in` and `:out` are acceptable values).
-
-
-### Performance
-This implementation is designed to perform well on large graphs. There are
-implementations which are marginally faster in practice for smaller graphs,
-but the performance improvements using this implementation on large graphs
-can be significant.
+    Traverse a graph `g` starting at vertex `s` / vertices `ss` using algorithm `alg`, maintaining state in [`AbstractTraversalState`](@ref) `state`. Next vertices to be visited are determined by `neighborfn` (default `outneighbors`). Return `true` if traversal finished; `false` if one of the visit functions caused an early termination.
+    
 """
-bfs_parents(g::AbstractGraph, s::Integer; dir = :out) = 
-    (dir == :out) ? _bfs_parents(g, s, outneighbors) : _bfs_parents(g, s, inneighbors)
+function traverse_graph!(
+    g::AbstractGraph{U},
+    ss::AbstractVector,
+    alg::BFS,
+    state::AbstractTraversalState,
+    neighborfn::Function=outneighbors
+    ) where U<:Integer
 
-function _bfs_parents(g::AbstractGraph{T}, source, neighborfn::Function) where T
+
     n = nv(g)
     visited = falses(n)
-    parents = zeros(T, nv(g))
-    cur_level = Vector{T}()
+    cur_level = Vector{U}()
     sizehint!(cur_level, n)
-    next_level = Vector{T}()
+    next_level = Vector{U}()
     sizehint!(next_level, n)
-    @inbounds for s in source
-        visited[s] = true
-        push!(cur_level, s)
-        parents[s] = s
+    @inbounds for s in ss
+        us = U(s)
+        visited[us] = true
+        push!(cur_level, us)
+        initfn!(state, us) || return false
     end
     while !isempty(cur_level)
         @inbounds for v in cur_level
-            @inbounds @simd for i in  neighborfn(g, v)
+            previsitfn!(state, v) || return false
+            @inbounds @simd for i in neighborfn(g, v)
+                visitfn!(state, v, i) || return false
                 if !visited[i]
+                    newvisitfn!(state, v, i) || return false
                     push!(next_level, i)
-                    parents[i] = v
                     visited[i] = true
                 end
             end
+            postvisitfn!(state, v) || return false
         end
+        postlevelfn!(state) || return false
         empty!(cur_level)
         cur_level, next_level = next_level, cur_level
-        sort!(cur_level)
+        sort!(cur_level, alg=alg.sort_alg)
     end
-    return parents
+    return true
+end
+
+mutable struct DistanceState{T<:Integer} <: AbstractTraversalState
+    distances::Vector{T}
+    n_level::T
+end
+
+@inline function initfn!(s::DistanceState{T}, u) where T 
+    s.distances[u] = zero(T)
+    return true
+end
+@inline function newvisitfn!(s::DistanceState, u, v) 
+    s.distances[v] = s.n_level
+    return true
+end
+@inline function postlevelfn!(s::DistanceState{T}) where T
+    s.n_level += one(T)
+    return true
 end
 
 """
-    bfs_tree(g, s[; dir=:out])
+    distances(g, s, alg=BFS())
+    distances(g, ss, alg=BFS())
 
-Provide a breadth-first traversal of the graph `g` starting with source vertex `s`,
-and return a directed acyclic graph of vertices in the order they were discovered.
-If `dir` is specified, use the corresponding edge direction (`:in` and `:out` are
-acceptable values).
+Return a vector filled with the geodesic distances of vertices in  `g` from vertex `s` / unique vertices `ss`
+using BFS traversal algorithm `alg`.  For vertices in disconnected components the default distance is `typemax(T)`.
 """
-bfs_tree(g::AbstractGraph, s::Integer; dir = :out) = tree(bfs_parents(g, s; dir = dir))
-
-"""
-    gdistances!(g, source, dists; sort_alg=QuickSort)
-
-Fill `dists` with the geodesic distances of vertices in `g` from source vertex (or
-collection of vertices) `source`. `dists` should be a vector of length `nv(g)` 
-filled with `typemax(T)`. Return `dists`.
-
-For vertices in disconnected components the default distance is `typemax(T)`.
-
-An optional sorting algorithm may be specified (see Performance section).
-
-### Performance
-`gdistances` uses `QuickSort` internally for its default sorting algorithm, since it performs
-the best of the algorithms built into Julia Base. However, passing a `RadixSort` (available via
-[SortingAlgorithms.jl](https://github.com/JuliaCollections/SortingAlgorithms.jl)) will provide
-significant performance improvements on larger graphs.
-"""
-function gdistances!(g::AbstractGraph{T}, source, vert_level; sort_alg = QuickSort) where T
-    n = nv(g)
-    visited = falses(n)
-    n_level = one(T)
-    cur_level = Vector{T}()
-    sizehint!(cur_level, n)
-    next_level = Vector{T}()
-    sizehint!(next_level, n)
-    @inbounds for s in source
-        vert_level[s] = zero(T)
-        visited[s] = true
-        push!(cur_level, s)
-    end
-    while !isempty(cur_level)
-        @inbounds for v in cur_level
-            @inbounds @simd for i in outneighbors(g, v)
-                if !visited[i]
-                    push!(next_level, i)
-                    vert_level[i] = n_level
-                    visited[i] = true
-                end
-            end
-        end
-        n_level += one(T)
-        empty!(cur_level)
-        cur_level, next_level = next_level, cur_level
-        sort!(cur_level, alg = sort_alg)
-    end
-    return vert_level
-end
-
-"""
-    gdistances(g, source; sort_alg=QuickSort)
-
-Return a vector filled with the geodesic distances of vertices in  `g` from
-`source`. If `source` is a collection of vertices each element should be unique.
-For vertices in disconnected components the default distance is `typemax(T)`.
-
-An optional sorting algorithm may be specified (see Performance section).
-
-### Performance
-`gdistances` uses `QuickSort` internally for its default sorting algorithm, since it performs
-the best of the algorithms built into Julia Base. However, passing a `RadixSort` (available via
-[SortingAlgorithms.jl](https://github.com/JuliaCollections/SortingAlgorithms.jl)) will provide
-significant performance improvements on larger graphs.
-"""
-gdistances(g::AbstractGraph{T}, source; sort_alg = Base.Sort.QuickSort) where T = gdistances!(g, source, fill(typemax(T), nv(g)); sort_alg = sort_alg)
-
-"""
-    has_path(g::AbstractGraph, u, v; exclude_vertices=Vector())
-
-Return `true` if there is a path from `u` to `v` in `g` (while avoiding vertices in
-`exclude_vertices`) or `u == v`. Return false if there is no such path or if `u` or `v`
-is in `excluded_vertices`. 
-"""
-function has_path(g::AbstractGraph{T}, u::Integer, v::Integer; 
-        exclude_vertices::AbstractVector = Vector{T}()) where T
-    seen = zeros(Bool, nv(g))
-    for ve in exclude_vertices # mark excluded vertices as seen
-        seen[ve] = true
-    end
-    (seen[u] || seen[v]) && return false
-    u == v && return true # cannot be separated
-    next = Vector{T}()
-    push!(next, u)
-    seen[u] = true
-    while !isempty(next)
-        src = popfirst!(next) # get new element from queue
-        for vertex in outneighbors(g, src)
-            vertex == v && return true
-            if !seen[vertex]
-                push!(next, vertex) # push onto queue
-                seen[vertex] = true
-            end
-        end
-    end
-    return false
+function distances(g::AbstractGraph{T}, s, alg::BFS=BFS()) where T
+    d = fill(typemax(T), nv(g))
+    state = DistanceState(d, one(T))
+    traverse_graph!(g, s, alg, state)
+    return state.distances
 end
